Move SearchServer into its own files.
